🚧 9T0RT3 task: split retrieval query planner
