+++ .agentplane/tasks/202607221852-9T0RT3/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221852-9T0RT3/build-deterministic-task-knowledge-retrieval",
   "created_at": "2026-07-30T10:59:23.993Z",
-  "diffstat_sha256": "sha256:2f16ec47c3d776aae7b748ea2dbb75f90abaa8223a3732c28a03e97e9182ff20",
+  "diffstat_sha256": "sha256:cdd13a3d8488646ddfeb50b2e1c5ba68ba9755a881ad77c18bf5f6947e767a7b",
   "last_verified_at": "2026-07-30T11:52:33.137Z",
   "last_verified_diffstat_sha256": "sha256:25838848e55bdd819bf7cc8cad3ab17042d18fb5cf07b6d74df4d39ffe5fb25e",
   "pr_number": 4690,
